Multiple Spreadsheet support added, Font Upgraded
Support for using multiple spreadsheets has been added, will work where
the tilesize is 1, 2, 4, 8, 16, 32, 64, ...  Font has also been upgraded
because of this so that different fonts can be used based on a
spreadsheet, more work will have to be done though because it relies on
the sizing of the current spritesheet.  Different sized sheets need to
be tested on this yet and the spritesheets need to be split up into
different images.