defaultTiles setup
Added defaultFont.psd to go with the PNG.  Created a defaultTiles
spritesheet(PSD and PNG) and changed the coding so that it could now be
used.  This sheet is 16 tiles wide, whereas the old one was 32, so we
know that different(on x) sized tile grids will work fine.  Now testing
is just needed with different tile sizes.  SpriteSheet now requires the
tileSize to be put into the constructor, the old constructor was pulled
out.