Tilesets added
Tilesets setup(with a default tileset) so that each level can have a
different tileset associated with it.  Tilesets will need more work on
it as things develop to provide this kind of functionality, but it will
be a way to have different block types defined with different properties
and colors(to go along with a different spritesheet, if wanted).  The
level class will have to be updated to handle the list of tiles and
entities in it, probably a linked list the traverses them all with a
tick function on the head.  the current way that it is setup, it will
not work.  it may be that the Tile class should be split out into a
class that will handle the properties (the way it behaves and looks -
basically the current system) and another class that will handle the
state that it's in.  Something similar will have to be setup for the
entities in a level (probably with separate lists of players, hostile
mobs, items, ...).